Pin mime-types at < 2.0.0 on rubies < 1.9
Thinking that removing support for 1.8 since it's been EOL'd would
be the sane choice here, but I am not entirely certain that I want
to make that choice all up at front.

The easiest thing to do, from a community perspective would be to
remove support for 1.8 in Savon 3 and to bump Wasabi, Nori, etc
at the same time.