Basically revert pull request #15
@paracycle was kind enough to track down these changes. I believe these changes
may be valid for Output, but almost certainly not for Input.

I had an opportunity to speak with @rubiii earlier about this and his thoughts
were that it wasn't very surprising that this change broke a lot of real code.
For one, it also changed specs. Secondly, there wasn't a real example of why this
would need to be like this in the PR. Lastly, the end result of the code is
overly complicated. I've just spent the last hour looking at the `input_output_for`
method in Wasabi::Parser just to find out that this was pulling in a completely
bogus request / response tag name in two thirds of the cases.

My hopes are that this fix will fix the regression in Savon 2.3.x and we can
stop worrying about this for Savon v3.