Add specs to cover failing cases
Hoping to see PR #39 fix the one failing test here

The real crazy thing here is that we did not have specs that covered both
cases. Sometimes you expect part of the operation message to be changed, but
sometimes you really expect that it will Do The Right Thing™ and not bork
your operation name.

A huge shout-out and thank you to @leoping for helping track this down
and for actually fixing the bug.