Make tests in PR #46 pass
There's a lot of convoluted logic going on in Wasabi for determining the
"correct name" for an operation. Around this time in 2013, a change in PR #15
that made the parser fall back to the port message's message attribute.

I searched around in the W3 spec for where this is the correct behavior and all
I could find was hints that the operation name should always be used.

Unsure if and or when this will ever be determined correctly, but this seems
to be a step in the right direction.