Revert "Add support for rpc encoded wsdl"
This change breaks backwards compatibility, by changing the parsing of
the SOAP messages.

I do not know if the change is incorrect. But if introduced, according
to SemVer it must be introduced in a new major release, to give
downstreams the chance to port their code accordingly.

Fixes #99

See also https://github.com/savonrb/savon/issues/935

This reverts commit 6fc97c6048b28aa79d053dc7bf9b736513c42c11.